Don't crash on some malformed cue file lines
Also add a related test. I saw something related in crash reports. It was
during the .toLong() however, so I'm not completely sure if this fixes the
issue observed there, too, but I couldn't come up with any way of causing a
backtrace identical to the one in the crash report.